// client/src/utilities/tests/dataHooks.ts
export enum DataHooks {
  landingPageGetStartedButton = "landingPageRegisterButton",
  landingPageHeader = "landingPageHeader",
  landingPageDescription = "landingPageDescription",
  registerPageCopyrights = "registerPageCopyrights",
  registerPageFirstName = "registerPageFirstName",
  registerPageLastName = "registerPageLastName",
  registerPageEmail = "registerPageEmail",
  registerPagePassword = "registerPagePassword",
  registerPagePasswordConfirm = "registerPagePasswordConfirm",
  registerPageSignUp = "registerPageSignUp",
  registerPageSignUpButton = "registerPageSignUpButton",
  registerPageSignInLink = "registerSignInLink",
}